Pause Menu Implemented
Adds functions of restarting, going back to title screen, and going back
to the checkpoint. Fixed the problem with the screen changing colors
after unpausing. The pause menu though is positioned in one place, due
the menu being put in the center of the level. Since we only use the
corner, this is to be expected. Will change once we implement the
dynamic camera for next iteration.